+++ watch/mlops/manager.py
@@ -1001,6 +1001,8 @@ def checkpoint_filepath_info(fname):
     # We assume it must have this
     suffix = ''.join(fname.partition('epoch')[1:])
     # Hack: making name assumptions
+    # could probably parse the right one out of filename in lightning monitor
+    # callbacks if we get access to that sort of into.
     parsers = [
         parse.Parser('epoch={epoch:d}-step={step:d}-{ckpt_ver}.{ext}'),
         parse.Parser('epoch={epoch:d}-step={step:d}.{ext}'),
@@ -1017,6 +1019,9 @@ def checkpoint_filepath_info(fname):
         parse.Parser('epoch={epoch:d}-val_loss={val_loss:f}'),
         parse.Parser('epoch={epoch:d}-val_loss={val_loss:f}.{ext}'),
         parse.Parser('epoch={epoch:d}-val_loss={val_loss:f}.{ext1}.{ext}'),
+        #
+        parse.Parser('epoch={epoch:d}-step={step:d}-val_loss={val_loss:f}'),
+        parse.Parser('epoch={epoch:d}-step={step:d}-val_loss={val_loss:f}.{ext1}.{ext}'),
     ]
     # results = parser.parse(str(path))
     info = None