+++ CHANGELOG.md
@@ -4,7 +4,16 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
-## Version 0.7.3 - Target 2023-06-xx
+## Version 0.8.0 - Target 2023-07-xx
+
+### Fixed
+* Fixed issue with kwcoco directory roots in prepare splits.
+
+### Removed
+* Removed old fit script. The lightning CLI script is considered stable.
+
+
+## Version 0.7.5 - Target 2023-06-30
 
 ### Added
 * Add transfer-feature script to COLD task to port features to time-averaged data